requirements.txt
Pillow==9.5.0
pystray==0.19.4
